Разбил main и исправил ошибки

+++ src/main/java/org/Dubrovsky/Aston/Homework/third/AdapterAndStrategy/APIs/PaypalApi.java
@@ -1,8 +1,10 @@
 package org.Dubrovsky.Aston.Homework.third.AdapterAndStrategy.APIs;
 
+import java.math.BigDecimal;
+
 public class PaypalApi {
 
-    public void sendMoney(String currency, double amount) {
+    public void sendMoney(String currency, BigDecimal amount) {
         System.out.println("PayPal: отправлено " + amount + " " + currency);
     }
 }
